+++ packages/core/src/entity/decorator/index.ts
@@ -1,6 +1,7 @@
-export { getEntities, getMeta } from './definition.js';
+export { defineHook, getEntities, getMeta } from './definition.js';
 export * from './entity.js';
 export * from './field.js';
+export * from './hook.js';
 export * from './id.js';
 export * from './index-decorator.js';
 export * from './relation.js';