+++ packages/core/src/util/index.ts
@@ -1,5 +1,6 @@
 export * from './dialect.util.js';
 export * from './field.util.js';
+export * from './hook.util.js';
 export * from './logger.js';
 export * from './object.util.js';
 export * from './raw.js';